+++ CHANGELOG.md
@@ -125,6 +125,7 @@ high state of flux, you're at risk of it changing without notice.
     - add `Show` instance (@gcanti)
   - `BooleanAlgebra`
     - deprecate `booleanAlgebraBoolean`, use `boolean.BooleanAlgebra` instead (@gcanti)
+    - deprecate `getFunctionBooleanAlgebra, use`function.getBooleanAlgebra` instead (@gcanti)
   - `Compactable`
     - add `compact` helper (@gcanti)
     - add `separate` helper (@gcanti)